qubes-i3-sensible-terminal: use qrexec-client directly
Currently qvm-run is used to start a terminal in a vm. However this will
keep stdin/stdout connected depleting inter-VM shared memory which is
a limited resource (as pointed out by Marek Marczykowski-Górecki).

To avoid this, use qrexec-client directly with "-e" option to avoid
connecting stdin/stdout. Note that qvm-run does not support this option
at the moment, but it could be worked around by pointing stdin/stdout to
/dev/null. However using qrexec-cliend seems like a more efficient
solution in this case.

This should be fine since a running vm is already required to be able to
use qubes-i3-sensible-terminal.

QubesOS/qubes-issues#5969
Signed-off-by: Lukas Czerner <lukas@czerner.cz>