Fix starting notification daemon
When multiple notification daemons are installed, bus activation doesn't
work. So, just installing dunst is not enough, it needs to be started
explicitly. And for the same reason, systemd unit cannot be used,
because systemd refuses to load two units with the same bus name (even
if only one is started at the same time).

https://github.com/QubesOS/qubes-issues/issues/8768